% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonemes.R
\name{get_jtrace_language}
\alias{get_jtrace_language}
\title{Create jTRACE language (phonemic inventory)}
\usage{
get_jtrace_language(
  phonemes,
  features,
  language_name = NULL,
  output_path = NULL
)
}
\arguments{
\item{phonemes}{Character vector of length N with the phonemes to be implemented (in jTRACE notation)}

\item{features}{Data frame or matrix with N rows (phonemes) and 9 columns (acoustic features). Acoustic features are in the following order: burst (bur), voicing (voi), consonantal (con), acuteness (grd), diffuseness (dif), vocalic (voc), power (pow). See McClelland & Elman (1986).}

\item{language_name}{Name of the language that will be created}

\item{duration_scalar}{Numeric vector of length Duration scalar}

\item{allphonic_relations}{Allophonic relations}
}
\description{
Create jTRACE language (phonemic inventory)
}
