% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{ipa_to_jtrace}
\alias{ipa_to_jtrace}
\title{Transcribe phonology from IPA to jTRACE notation}
\usage{
ipa_to_jtrace(x, keep_other = FALSE)
}
\arguments{
\item{x}{A character vector with the phonological forms to be transcribed}

\item{keep_other}{Should symbols other than phonemes be kept in the
transcriptions? Defaults to FALSE}
}
\value{
A character vector with the jTRACE transcriptions of the provided
phonological forms
}
\description{
Transcribe phonology from IPA to jTRACE notation
}
\details{
\enumerate{
\item If \code{keep_other}, special characters (symbols that do not
correspond to phonemes in the \code{phonemes} data set, such as apostrophes
or dots) are removed. 2) Colons (:) are replaced with the previous symbol,
since they are interpreted as a modifier of the duration of the previous
phoneme. 3) Pairwise replacements are performed according to the
\code{phonemes} data set.
}
}
