% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonemes.R
\name{jtrace_create_language}
\alias{jtrace_create_language}
\title{Create jTRACE language (phonemic inventory)}
\usage{
jtrace_create_language(
  phonemes = NULL,
  features,
  duration_scalar = NULL,
  allophonic_relations = NULL,
  language_name
)
}
\arguments{
\item{phonemes}{Character vector indicating the jTRACE notation of each phoneme. It must be the same length as the number of rows of the matrix or data frame introduced in \code{features}. This argument can be left NULL (default) if the matrix or data frame introduced in \code{features} has appropriate row names indicating the jTRACE notation of the phonemes.}

\item{features}{A M x N matrix or data frame (where M is the number of phonemes and N is 7, the number of features) that contains the values of the features (columns) for each phoneme (rows) with a score ranging from 0 to 9.}

\item{duration_scalar}{Matrix or data frame indicating the values of the duration scalar, with each phoneme as a row and each feature as a column. If NULL (default), all duration values are set to 1.}

\item{allophonic_relations}{Array or data frame with logical values indicating whether each combination of phonemes is an allophone, with phonemes are rows and columns. If NULL (default), no allophonic relations are specified.}

\item{language_name}{Name of the language that will be created.}
}
\description{
Create jTRACE language (phonemic inventory)
}
\examples{
jtrace_create_language()
}
