% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{jtrace_create_lexicon}
\alias{jtrace_create_lexicon}
\title{Create jTRACE lexicon}
\usage{
jtrace_create_lexicon(phonology, frequency, lexicon_name)
}
\arguments{
\item{phonology}{Character vector with the jTRACE phonological transcription of the word forms}

\item{frequency}{Numeric vector with the lexical frequencies of the word forms}

\item{lexicon_name}{Character string indicating the name of the lexicon that will be generated}
}
\description{
Create jTRACE lexicon
}
\examples{
my_phons <- c("plEIn", "kEIk", "taIɡ@", "ham", "sit")
my_freqs <- c(0.0483, 0.0804, 0.0288, 0.0282, 0.0767)
jtrace_create_lexicon(phonology = my_phons, frequency = my_freqs, lexicon_name = "my_lex")
}
