% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{jtrace_get_frequency}
\alias{jtrace_get_frequency}
\title{Extract lexical frequencies}
\usage{
jtrace_get_frequency(word, language = "English", scale = "frequency_abs")
}
\arguments{
\item{word}{Character vector with the orthographic word form}

\item{language}{Character vector containing the language(s) to lookup the frequency of the words for. Must be one or more of "English", "Spanish, and/or "Catalan".}

\item{scale}{Character vector indicating the scale(s) of the frequency scores. Must be one or more of "frequency_abs" (absolute frequency), "frequency_rel", (relative frequency, \code{counts/1e6}, default), or "frequency_zipf" (\code{log10(counts*1e6)+3})"}
}
\value{
A data frame containing a column for the words and one column for the SUBTLEX frequencies in each language for the same word
}
\description{
Extract lexical frequencies
}
\examples{
\dontrun{
my_words <- c("plane", "cake", "tiger", "ham", "seat")
jtrace_get_frequency(word = my_words, language = "English", scale = "frequency_abs")
jtrace_get_frequency(word = my_words, language = c("Spanish", "Catalan"), scale = "frequency_zipf")
jtrace_get_frequency(word = my_words, language = c("Spanish"), scale = "frequency_rel")
}
}
\references{
\describe{
\item{English}{Van Heuven, W. J., Mandera, P., Keuleers, E., & Brysbaert, M. (2014). SUBTLEX-UK: A new and improved word frequency database for British English. Quarterly journal of experimental psychology, 67(6), 1176-1190.}
\item{Spanish}{Cuetos, F., Glez-Nosti, M., Barbon, A., & Brysbaert, M. (2011). SUBTLEX-ESP: frecuencias de las palabras espanolas basadas en los subtitulos de las peliculas. Psicológica, 32(2), 133-144.}
\item{Catalan}{Boada, R., Guasch, M., Haro, J., Demestre, J., & Ferré, P. (2020). SUBTLEX-CAT: Subtitle word frequencies and contextual diversity for Catalan. Behavior research methods, 52(1), 360-375.}
}
}
\author{
Gonzalo Garcia-Castro \href{mailto:gonzalo.garciadecastro@upf.edu}{gonzalo.garciadecastro@upf.edu}
}
