% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonemes.R
\name{jtrace_get_language}
\alias{jtrace_get_language}
\title{Get jTRACE language}
\usage{
jtrace_get_language(language_name = NULL)
}
\arguments{
\item{language_name}{Character vector of length 1 indicating the jTRACE language to import}
}
\description{
Get jTRACE language
}
