% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{jtrace_get_lexicon}
\alias{jtrace_get_lexicon}
\title{Get jTRACE lexicon}
\usage{
jtrace_get_lexicon(lexicon = NULL)
}
\arguments{
\item{lexicon}{Character vector of length 1 indicating the jTRACE lexicon to import}
}
\value{
A data frame with the contents of the retrieved jTRACE lexicon.
This data frame will always contain a column for word forms (\code{phonology}).
Sometimes, it will also contain another column for lexical frequencies (\code{frquency}),
depending on the information provided in that file in the jTRACE original implementation.
}
\description{
Get jTRACE lexicon
}
\examples{
jtrace_get_lexicon("sevenlex")
}
