% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{jtrace_launch}
\alias{jtrace_launch}
\title{Launches jTRACE}
\usage{
jtrace_launch()
}
\description{
Launches jTRACE
}
\examples{
\donttest{jtrace_launch()}
}
\references{
Strauss, T. J., Harris, H. D., & Magnuson, J. S. (2007). jTRACE: A reimplementation and extension of the TRACE model of speech perception and spoken word recognition. Behavior Research Methods, 39(1), 19-30.
}
\author{
Gonzalo Garcia-Castro \href{mailto:gonzalo.garciadecastro@upf.edu}{gonzalo.garciadecastro@upf.edu}
}
