% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonemes.R
\name{jtrace_list_languages}
\alias{jtrace_list_languages}
\title{List jTRACE available languages}
\usage{
jtrace_list_languages()
}
\value{
A character vector listing the available languages in the jTRACE folder
}
\description{
List jTRACE available languages
}
\examples{
jtrace_list_languages()
}
