% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{jtrace_list_lexicons}
\alias{jtrace_list_lexicons}
\title{List jTRACE available lexicons}
\usage{
jtrace_list_lexicons()
}
\value{
A character vector listing the available lexicons in the jTRACE folder
}
\description{
List jTRACE available lexicons
}
\examples{
\dontrun{jtrace_list_lexicons()}
}
\references{
Strauss, T. J., Harris, H. D., & Magnuson, J. S. (2007). jTRACE: A reimplementation and extension of the TRACE model of speech perception and spoken word recognition. Behavior Research Methods, 39(1), 19-30.
}
\seealso{
\code{\link{jtrace_get_lexicon}} for importing a lexicon, and \code{\link{jtrace_create_language}} for creating a new lexicon.
}
\author{
Gonzalo Garcia-Castro \href{mailto:gonzalo.garciadecastro@upf.edu}{gonzalo.garciadecastro@upf.edu}
}
