% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonemes.R
\docType{data}
\name{phonemes}
\alias{phonemes}
\title{Phoneme inventory}
\format{
A data frame with 74 rows and 7 variables:
\describe{
\item{id}{Individual phoneme numeric identifier}
\item{ipa}{Phoneme symbol in IPA notation}
\item{approximation}{Closest IPA phoneme this phoneme will be collapsed into}
\item{sampa}{Phoneme symbol in SAMPA notation}
\item{trace}{Phoneme in jTRACE notation},
\item{description}{Rough classification of the phoneme}
\item{type}{Is the phoneme a consonant or a vowel?}
}
}
\usage{
data(phonemes)
}
\description{
A dataset containing an extensive set of phonemes that can be transcribed from IPA or SAMPA to jTRACE notation using approximations.
}
\references{


Strauss, T. J., Harris, H. D., & Magnuson, J. S. (2007). jTRACE: A reimplementation and extension of the TRACE model of speech perception and spoken word recognition. Behavior Research Methods, 39(1), 19-30.
}
\keyword{datasets}
