% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{set_jtrace_path}
\alias{set_jtrace_path}
\title{Set jRTACE path}
\usage{
set_jtrace_path(path = NULL)
}
\arguments{
\item{path}{Character string indicating the path in which to install jTRACE}
}
\description{
Set jRTACE path
}
