% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{jtrace_check_java}
\alias{jtrace_check_java}
\title{Check if Java is installed}
\usage{
jtrace_check_java()
}
\value{
TRUE if Java is installed and version is 1.4 or higher, FALSE otherwise
}
\description{
Check if Java is installed
}
\details{
jTRACE requires Java 1.4 or higher. This function checks if Java installed, and if the current
version is recent enough for Java to run. If not, you will be pointed to the Java
website where you can download the most recent version of Java. After installing it, you will
be able to install and launch jTRACE.
}
\examples{
jtrace_check_java()
}
\references{
Strauss, T. J., Harris, H. D., & Magnuson, J. S. (2007). jTRACE: A reimplementation and extension of the TRACE model of speech perception and spoken word recognition. Behavior Research Methods, 39(1), 19-30.
}
\author{
Gonzalo Garcia-Castro \href{mailto:gonzalo.garciadecastro@upf.edu}{gonzalo.garciadecastro@upf.edu}
}
