% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{jtrace_get_frequency}
\alias{jtrace_get_frequency}
\title{Extract lexical frequencies}
\usage{
jtrace_get_frequency(word, language = "English", scale = "frequency_rel")
}
\arguments{
\item{word}{Character vector with the orthographic word form}

\item{language}{Character vector containing the language(s) to lookup the frequency of the words for. Must be one or more of "English", "Spanish, and/or "Catalan".}

\item{scale}{Character vector indicating the scale(s) of the frequency scores. Must be one or more of "frequency_abs" (absolute frequency), "frequency_rel", (relative frequency, \code{counts/1e6}, default), or "frequency_zipf" (\code{log10(counts*1e6)+3})"}
}
\value{
A data frame containing a column for the words and one column for the SUBTLEX frequencies in each language for the same word
}
\description{
Extract lexical frequencies
}
\examples{
my_words <- c("plane", "cake", "tiger", "ham", "seat")
jtrace_get_frequency(words = my_words, language = "English", scale = "frequency_rel")
jtrace_get_frequency(words = my_words, language = c("Spanish", "Catalan"), scale = "frequency_zipf")
jtrace_get_frequency(words = my_words, language = c("Spanish"), scale = "frequency_abs")
}
