% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{jtrace_get_language}
\alias{jtrace_get_language}
\title{Get jTRACE language}
\usage{
jtrace_get_language(language_name = "default")
}
\arguments{
\item{language_name}{Character vector of length 1 indicating the jTRACE language to import. Defaults to "default".}
}
\value{
A list of data frames containing a data frame for the phonemes and their
scores across the seven features implemented in jTRACE (\code{features}),
a data frame containing the duration scalars of each phoneme for the even features
implemented in jTRACE (\code{duration_scalar}), and a data frame containing the
allophonic relations between the phonemes (\code{allophonic_relations}).
@examples
jtrace_get_language("default")
}
\description{
Get jTRACE language
}
\references{
Strauss, T. J., Harris, H. D., & Magnuson, J. S. (2007). jTRACE: A reimplementation and extension of the TRACE model of speech perception and spoken word recognition. Behavior Research Methods, 39(1), 19-30.
}
\seealso{
\code{\link{jtrace_list_languages}} for listing available languages, and \code{\link{jtrace_create_language}} for creating a new language.
}
\author{
Gonzalo Garcia-Castro \href{mailto:gonzalo.garciadecastro@upf.edu}{gonzalo.garciadecastro@upf.edu}
}
