% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{jtrace_get_lexicon}
\alias{jtrace_get_lexicon}
\title{Get jTRACE lexicon}
\usage{
jtrace_get_lexicon(lexicon = NULL)
}
\arguments{
\item{lexicon}{Character vector of length 1 indicating the jTRACE lexicon to import}
}
\description{
Get jTRACE lexicon
}
