% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{jtrace_install}
\alias{jtrace_install}
\title{Download and install jTRACE}
\usage{
jtrace_install(path = NULL, overwrite = NULL)
}
\arguments{
\item{path}{Character string indicating the path in which to install jTRACE}

\item{overwrite}{Logical value indicating whether to replace an existing jTRACE folder, in case there is}
}
\description{
Download and install jTRACE
}
