% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{jtrace_install}
\alias{jtrace_install}
\title{Download and install jTRACE}
\usage{
jtrace_install(overwrite = NULL, quiet = FALSE)
}
\arguments{
\item{overwrite}{Logical value indicating whether to replace an existing jTRACE folder, in case there is}

\item{quiet}{Should downloading progress not be shown?}
}
\description{
Download and install jTRACE
}
\examples{
jtrace_install()
}
\author{
Gonzalo Garcia-Castro \href{mailto:gonzalo.garciadecastro@upf.edu}{gonzalo.garciadecastro@upf.edu}
}
