% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{jtrace_list_languages}
\alias{jtrace_list_languages}
\title{List jTRACE available languages}
\usage{
jtrace_list_languages()
}
\value{
A character vector listing the available languages in the jTRACE folder
}
\description{
List jTRACE available languages
}
\examples{
jtrace_list_languages()

}
\references{
Strauss, T. J., Harris, H. D., & Magnuson, J. S. (2007). jTRACE: A reimplementation and extension of the TRACE model of speech perception and spoken word recognition. Behavior Research Methods, 39(1), 19-30.
}
\seealso{
\code{\link{jtrace_get_language}} for importing a language, and \code{\link{jtrace_create_language}} for creating a new language.
}
\author{
Gonzalo Garcia-Castro \href{mailto:gonzalo.garciadecastro@upf.edu}{gonzalo.garciadecastro@upf.edu}
}
