% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{jtrace_list_lexicons}
\alias{jtrace_list_lexicons}
\title{List jTRACE available lexicons}
\usage{
jtrace_list_lexicons()
}
\value{
A character vector listing the available lexicons in the jTRACE folder
}
\description{
List jTRACE available lexicons
}
\examples{
jtrace_list_lexicons()
}
